Got access to the logged in user's timeline pictures (not their friend's pictures) and got them posted. Had to see the raw JSON response and then parse necessary areas. Also created many conditions in order to store the right values into the database.
Created a time converter specifically from Facebook (RFC 822 to Epoch Unix Timestamp) and it has to be formatted in a certain way for Java to interpret it right (raw JSON date was not good enough and gave ParseException). --> ended up not using the SimpleFacebook photo code due not being able to capture information.

Everything successfully inserts into the database with no strange values.

Issue:
No square profile image options from Facebook JSON data.
Also too much work is done on the main thread. Need to correct this to make it efficient.
As of now, user is required to sign in twice. Need to find the line responsible for this and make the user only sign in once.