Modified the current PostItem model to include the OP's profile id as part of the object. Made any modifications necessary to any of the databases to insert this information into the table. (No relational database stuff yet)
Modified the User object to add the user id to it for future callings.

Created layout for the the collection mode dialog. Added necessary function such as recyclerview and add user button. Eventually a dialog will be created. Changed the button for positive and negative (text color only)

Modified the addUserDialog and added a newInstance() method.