Added input counter to the collection naming input. Also improved the coding of the counter and adjust it to whatever character limit I want by just changing one variable value.
Readjusted the layout for the counter portion to conform to the above.

Added the necessary xml code for the sliding tab layout in the addCollectionDialog but the view is not visible yet and in progress.
Added two classes that will allow the implementation of a sliding tab layout

Imported support class v13 to allow usage of getFragmentManager() for the FragmentPageAdapter class to be used appropriately rather than using v4 by default.